Widen speed limit types from uint16 to uint32
The upload/download speed limits (s_maxupload, s_maxdownload) and slot
allocation (s_slotallocation) were stored as uint16, capping configurable
speed to 65534 kB/s (~524 Mbps) since 65535 is treated as unlimited.
This is inadequate for modern gigabit+ connections.

Changes:

Preferences.h / Preferences.cpp:
- Change s_maxupload, s_maxdownload, s_slotallocation from uint16 to
  uint32, along with their getters and setters
- The existing backwards compatibility check in CheckUlDlRatio() that
  maps 0xFFFF to UNLIMITED still works correctly with uint32

UploadQueue.h / UploadQueue.cpp:
- Change GetMaxSlots() return type and local variables from uint16 to
  uint32 to prevent truncation when calculating slot counts from large
  speed values

RemoteConnect.h:
- Update SetPreferencesConnections() parameter types for
  MaxDownloadSpeed, MaxUploadSpeed, and UploadSlotAllocation from
  uint16 to uint32

EC protocol note:
- EC tags EC_TAG_CONN_MAX_UL, EC_TAG_CONN_MAX_DL, and
  EC_TAG_CONN_SLOT_ALLOCATION now carry uint32 values. EC clients
  reading these tags should use GetInt() which handles any integer
  size. Clients sending these tags should encode them as 32-bit.

+++ src/UploadQueue.cpp
@@ -294,16 +294,16 @@ void CUploadQueue::Process()
 }
 
 
-uint16 CUploadQueue::GetMaxSlots() const
+uint32 CUploadQueue::GetMaxSlots() const
 {
-	uint16 nMaxSlots = 0;
+	uint32 nMaxSlots = 0;
 	float kBpsUpPerClient = (float)thePrefs::GetSlotAllocation();
 	if (thePrefs::GetMaxUpload() == UNLIMITED) {
 		float kBpsUp = theStats::GetUploadRate() / 1024.0f;
-		nMaxSlots = (uint16)(kBpsUp / kBpsUpPerClient) + 2;
+		nMaxSlots = (uint32)(kBpsUp / kBpsUpPerClient) + 2;
 	} else {
 		if (thePrefs::GetMaxUpload() >= 10) {
-			nMaxSlots = (uint16)floor((float)thePrefs::GetMaxUpload() / kBpsUpPerClient + 0.5);
+			nMaxSlots = (uint32)floor((float)thePrefs::GetMaxUpload() / kBpsUpPerClient + 0.5);
 				// floor(x + 0.5) is a way of doing round(x) that works with gcc < 3 ...
 			if (nMaxSlots < MIN_UP_CLIENTS_ALLOWED) {
 				nMaxSlots=MIN_UP_CLIENTS_ALLOWED;
@@ -413,7 +413,7 @@ void CUploadQueue::AddClientToQueue(CUpDownClient* client)
 				//   have been kicked so a slot is free again)
 				// - or there is a free slot, which means there is no HighID client on queue
 				if (client->m_bAddNextConnect) {
-					uint16 maxSlots = GetMaxSlots();
+					uint32 maxSlots = GetMaxSlots();
 					if (lastupslotHighID) {
 						maxSlots++;
 					}
